types: refactor fixing circular dependencies
